# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Neoting API
  version: 0.1.0-draft
  summary: Chat-first document-to-bookkeeping platform for UK accounting practices.
  description: |
    The single API. The public API is this same API with scoped OAuth clients —
    there is no second door (D10, Governance §3).

    Money is integer pence. Nothing changes state outside the Review → Approve
    path. Tenancy is row-level security, never the URL. Every extracted value
    carries its provenance class. Errors are RFC 7807 `problem+json` with a
    stable `NT-` code.

    The reasoning behind each of those, and the conventions a consumer needs, are
    in `packages/contracts/CLAUDE.md` and in the header comment of
    `openapi.yaml`. They are deliberately not repeated here: this field is
    stamped into the header of every generated file.
  contact:
    name: Shakib Bin Kabir — tech lead
    url: https://github.com/Neovogent/neoting
  x-nt-governs:
    sourceOfTruth: docs/Source_Of_Truth.md v1.4 — §4 Stage 1-2, §8.2, §15
    governance: docs/Engineering_Governance.md v1.4 — §3, §5.2, §10, §13.4
    schema: prisma/schema.prisma
servers:
  - url: http://localhost:3000/v1
    description: |
      Local — docker compose. The canonical development environment.

      3000, not 3001. This is the port the API actually listens on and it is
      the same number in all four places that have to agree: `PORT` in
      `apps/api/src/config/env.ts`, `EXPOSE` in `apps/api/Dockerfile`,
...
security:
  - workspaceSession: []
...
paths:
...
  /documents/{documentId}/bank-match:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      operationId: getDocumentBankMatch
      summary: The document's bank match — suggested or confirmed
      description: |
        The one `matches` row that links this document to a bank transaction,
        with the transaction embedded: a CONFIRMED match when one exists, else
        the newest SUGGESTED one, else `match: null`. The suggested half is
        written by the automatic suggester (the same deterministic
        supplier + amount + date-window compare that auto-closes a chase);
        confirming it is a `bank.confirm-match` proposal through
        Review → Approve — this surface is read-only and a suggestion gates
        nothing (Governance §9.5).
      tags: [bank]
      x-nt-side-effect: none
      responses:
        '200':
          description: The match, or null when nothing links this document to a bank line.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentBankMatch' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /action-proposals:
    get:
      operationId: listActionProposals
      summary: List proposals — the approval queue
      description: |
        Every proposal the caller's RLS context can see, newest first. One
        endpoint serves the live approval queue (`state=CREATED&state=REVIEWED`)
        and the decided history — they differ only by filter, the same shape
        as `/documents` (METH Stage 12, issue #140).

        Reading the list is not reviewing: `[Approve]` stays unreachable until
        `POST .../review` has recorded what was actually rendered.
      tags: [action-proposals]
      x-nt-side-effect: none
      parameters:
        - $ref: '#/components/parameters/BusinessIdFilter'
        - name: state
          in: query
          description: Repeat to widen. Omitted means every state.
          required: false
          schema:
            type: array
            items: { $ref: '#/components/schemas/ProposalState' }
          style: form
          explode: true
        - name: kind
          in: query
          description: Repeat to widen. Omitted means every kind.
          required: false
          schema:
            type: array
            items: { $ref: '#/components/schemas/ProposalKind' }
          style: form
          explode: true
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of proposals, newest first.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema:
                type: object
                required: [data, pageInfo]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ActionProposal' }
                  pageInfo: { $ref: '#/components/schemas/PageInfo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createActionProposal
      summary: Propose a state change
      description: |
        The only way anything in the product changes state.

        `kind` must exist in the action-kind registry; unknown kinds are
        rejected outright. The payload is a discriminated union on `kind`, so the
        generated client types each action's payload exactly.

        This creates the proposal and **executes nothing**. The response carries
        no Approve affordance and the server will refuse an approval until
        `POST /action-proposals/{id}/review` has been called — enforced
        server-side, not in the UI (Governance §10.3). A chat model may call this
        endpoint; it can never call the approval one.
      tags: [action-proposals]
      x-nt-side-effect: proposal
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateActionProposalRequest' }
            examples:
              routeUnrouted:
                summary: Assign an unrouted document, and teach the router
                value:
                  kind: document.route
                  businessId: biz_burger
                  payload:
                    documentId: doc_031
                    inbox: COSTS
                    teachRouterForSender: true
              moveEntity:
                summary: Move a misrouted document between entities (SoT Stage 11)
                value:
                  kind: document.move-business
                  businessId: biz_cosmo
                  payload:
                    documentId: doc_022
                    toBusinessId: biz_burger
                    acknowledgedAddresseeMismatch: true
      responses:
        '201':
          description: |
            Proposal created in state `CREATED`. Not reviewed, not approved,
            nothing executed.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionProposal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: |
            The actor lacks permission to *propose* this kind. Permission to
            propose and permission to approve are checked independently
            (Governance §11.2).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422':
          description: |
            The kind is known but the payload is not executable — the document is
            in a state this action cannot apply to, or a referenced record is
            gone.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /bank-transactions:
    get:
      operationId: listBankTransactions
      summary: List bank transactions
      description: |
        The normalised feed (SoT Stage 7), newest first. Amounts are signed
        pence exactly as the feed records them. Confirming a suggested match
        is a `bank.confirm-match` proposal — nothing here mutates, and the
        unmatched set this returns is the same set chase detection reads,
        so the Bank screen and the chase list can never disagree.
      tags: [banking]
      x-nt-side-effect: none
      parameters:
        - $ref: '#/components/parameters/BusinessIdFilter'
        - name: accountId
          in: query
          description: Restrict to one bank account.
          required: false
          schema: { type: string }
        - name: matchState
          in: query
          description: Repeat to widen. Omitted means every state.
          required: false
          schema:
            type: array
            items: { $ref: '#/components/schemas/MatchState' }
          style: form
          explode: true
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of transactions, newest booked first.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema:
                type: object
                required: [data, pageInfo]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BankTransaction' }
                  pageInfo: { $ref: '#/components/schemas/PageInfo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ==========================================================================
  # BUSINESSES (METH Stage 2, issue #120)
  # ==========================================================================
...
components:
  securitySchemes:
    workspaceSession:
      type: apiKey
      in: cookie
      name: nt_session
      description: |
        The accountant workspace session. Device-bound, rotating refresh token,
        absolute lifetime ≤ 30 days, idle timeout ≤ 24 h (Governance §11.1).
        Every state-changing browser request is additionally CSRF-protected.

    portalSession:
      type: http
      scheme: bearer
      description: |
        A client's OTP portal session. Scoped to exactly the items the granting
        chase requested and nothing else — enforced by RLS via
        `app.session_scope = 'delegated_upload'` plus the granted item IDs
        (Governance §5.2), not by the handler.

        The link and OTP are deliberately forwardable to whoever physically holds
        the document; the audit trail records *requested-from* separately from
        *uploaded-by-delegated-session*.

    oauth2:
      type: oauth2
      description: |
        Public API clients (D10, 185-vote gap). The same API, scoped — there is
        no second door. `documents:write` grants proposal creation and never
        approval: approving is a human UI action by definition.
      flows:
        authorizationCode:
          authorizationUrl: https://api.neoting.neovogent.com/v1/oauth/authorize
          tokenUrl: https://api.neoting.neovogent.com/v1/oauth/token
          scopes:
            documents:read: Read documents, extractions and processing logs
            documents:write: Create documents and propose document actions
            proposals:read: Read action proposals

    stripeSignature:
      type: apiKey
      in: header
...
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from `pageInfo.nextCursor`. Never construct one by hand.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    BusinessIdFilter:
      name: businessId
      in: query
      required: false
      description: |
        Restrict to one client workspace. **Omit it to span every workspace the
        caller can see** — that is what makes cross-client questions a filter
        rather than a second endpoint.

        This narrows; it never widens. RLS decides the ceiling, so passing an id
        the caller cannot access returns an empty page or `404`, never someone
        else's data.
      schema: { type: string }
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUIDv4, unique per logical operation. Replays return the
        original result rather than acting twice. Reusing a key with a different
        payload is `409` — that is a client bug, and failing loudly beats
        silently doing the wrong thing to someone's books.
      schema: { type: string, format: uuid }
  headers:
    TraceId:
      description: |
        The trace id born at this entry point and carried through every service,
        job, model call and adapter call it causes. Quote it in a bug report and
        the whole journey is one Journey Inspector lookup.
      schema: { type: string }
    RateLimitLimit:
      description: Requests permitted in the current window.
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests left in the current window.
      schema: { type: integer }
    RateLimitReset:
      description: Seconds until the window resets.
      schema: { type: integer }
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: { type: integer }
  responses:
    BadRequest:
      description: The request failed schema validation at the boundary.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://neoting.com/problems/NT-VAL-001
            title: Validation failed
            status: 400
            code: NT-VAL-001
            detail: One or more fields are invalid.
            traceId: 01J8XN2ZQ5R7TQ9E3M4K6P
            errors:
              - field: byteSize
                message: must be a positive integer
    Unauthenticated:
      description: No valid session or token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: |
        Authenticated, but not permitted. Used for permission failures on
        resources the caller can see. A resource the caller cannot see returns
        `404` instead — the difference matters, because `403` would confirm the
        record exists.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: |
        No such record — or one the caller's RLS context does not reach. The two
        are deliberately indistinguishable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    IdempotencyConflict:
      description: This `Idempotency-Key` was already used with a different payload.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: Rate limit exceeded (Governance §11.8).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalError:
      description: |
        Unexpected. The response carries the trace id and nothing else — internal
        detail and PII never reach a client (Governance §3).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 problem detail. The only error shape this API returns.
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
          description: |
            Points at the runbook page for `code`. Governance §13.4 requires
            every code to have one before it passes review, so this link is never
            dead.
          examples: ['https://neoting.com/problems/NT-ING-004']
        title: { type: string, description: Short, human, stable for the code. }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail:
          type: string
          description: What went wrong for this occurrence, in plain English, never blaming the user.
        traceId: { type: string }
        errors:
          type: array
          description: Field-level detail on validation failures.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          type: [string, 'null']
          description: Pass back as `cursor`. Null when there is no next page.
        hasMore: { type: boolean }

    # ---- enums (verbatim from prisma/schema.prisma) ------------------------
...
    ProposalState:
      type: string
      enum: [CREATED, REVIEWED, APPROVED, EXECUTED, CANCELLED, EXPIRED]
...
    MatchState:
      type: string
      description: |
        Mirrors the Prisma `MatchState` enum exactly. `CONFIRMED` is reachable
        only through an approved `bank.confirm-match` proposal.
      enum: [UNMATCHED, SUGGESTED, CONFIRMED, EXCLUDED]
    MatchKind:
      type: string
      description: Mirrors the Prisma `MatchKind` enum exactly.
      enum: [EXACT, PROBABILISTIC, PARTIAL_PAYMENT, BATCH_PAYMENT, CREDIT_NOTE]
    DocumentBankMatch:
      type: object
      description: |
        `GET /documents/{documentId}/bank-match` — the link between one
        document and one bank line, or the honest absence of one.
      required: [match]
      properties:
        match:
          type: [object, 'null']
          required: [id, state, kind, confidence, matchedBy, transaction]
          properties:
            id: { type: string }
            state: { $ref: '#/components/schemas/MatchState' }
            kind: { $ref: '#/components/schemas/MatchKind' }
            confidence:
              type: [number, 'null']
              description: The suggesting engine's score. Triage only — it gates nothing (Governance §9.5).
            matchedBy:
              type: [string, 'null']
              description: '`auto-suggester` for machine suggestions; a confirmation records the approver on the proposal instead.'
            transaction: { $ref: '#/components/schemas/BankTransaction' }
...
    ProposalKind:
      type: string
      description: |
        The action-kind registry **for this surface**. It is an allow-list, not a
        suggestion: a chat context physically lacks tools outside its registered
        kinds (Governance §9.5), and an unknown kind is rejected with
        `NT-PRP-001`.

        METH Stage 2 (issue #120) added the four demo-push kinds — `chase.send`,
        `publish.batch`, `bank.confirm-match` and `rule.create` — registered as
        `ProposalNotImplementedError` stubs until their stages land executors.
        Later surfaces still add `rule.activate`, `policy.activate` and
        `setting.change`.

        **Open question for Shakib — deliberately not decided here.** The OTP
        portal's editable-overlay correction (SoT Stage 8.4) is a client fixing a
        misread number on their own just-uploaded receipt, in a car park, before
        the document has ever been reviewed. Putting Read review → Approve in
        front of that would be the wrong reading of §8.2, which governs the
        accountant workspace. But it touches the Review → Approve enforcement
        path, which is a stop-and-ask item, so this contract does not settle it —
        it lands with the portal surface in pass 3.
      enum:
        - document.route          # assign an unrouted document, optionally teaching the router
        - document.update-coding  # confirm or change supplier, category, totals, VAT treatment
        - document.move-business  # move between entities, with the addressee-mismatch warning
        - document.reprocess      # retry a failed extraction or ingest
        - document.reject         # reject with a visible reason
        - document.split          # re-split a batch by page range
        - document.archive        # archive or unarchive
        - chase.send              # send composed chase SMS; review shows every message verbatim
        - publish.batch           # publish reviewed documents to the ledger; review shows the server-computed preview
        - bank.confirm-match      # confirm a suggested document ↔ transaction match
        - rule.create             # create a coding rule; it activates only on approval
        - document.revoke-link    # revoke capability URLs already sitting in someone's ledger
        - business.offboard       # deactivate a client workspace (D32 slice); soft, books retained
    CreateActionProposalRequest:
      description: |
        A discriminated union on `kind`. The discriminator sits on the union
        itself — not on a wrapper — so the generated client narrows `payload` from
        `kind` alone, and an unknown `kind` fails at the boundary rather than
        reaching a handler that has to guess.
      oneOf:
        - $ref: '#/components/schemas/RouteProposalRequest'
        - $ref: '#/components/schemas/UpdateCodingProposalRequest'
        - $ref: '#/components/schemas/MoveBusinessProposalRequest'
        - $ref: '#/components/schemas/ReprocessProposalRequest'
        - $ref: '#/components/schemas/RejectProposalRequest'
        - $ref: '#/components/schemas/SplitProposalRequest'
        - $ref: '#/components/schemas/ArchiveProposalRequest'
        - $ref: '#/components/schemas/ChaseSendProposalRequest'
        - $ref: '#/components/schemas/PublishBatchProposalRequest'
        - $ref: '#/components/schemas/BankConfirmMatchProposalRequest'
        - $ref: '#/components/schemas/RuleCreateProposalRequest'
        - $ref: '#/components/schemas/RevokeLinkProposalRequest'
        - $ref: '#/components/schemas/BusinessOffboardProposalRequest'
      discriminator:
        propertyName: kind
        mapping:
          document.route: '#/components/schemas/RouteProposalRequest'
          document.update-coding: '#/components/schemas/UpdateCodingProposalRequest'
          document.move-business: '#/components/schemas/MoveBusinessProposalRequest'
          document.reprocess: '#/components/schemas/ReprocessProposalRequest'
          document.reject: '#/components/schemas/RejectProposalRequest'
          document.split: '#/components/schemas/SplitProposalRequest'
          document.archive: '#/components/schemas/ArchiveProposalRequest'
          chase.send: '#/components/schemas/ChaseSendProposalRequest'
          publish.batch: '#/components/schemas/PublishBatchProposalRequest'
          bank.confirm-match: '#/components/schemas/BankConfirmMatchProposalRequest'
          rule.create: '#/components/schemas/RuleCreateProposalRequest'
          document.revoke-link: '#/components/schemas/RevokeLinkProposalRequest'
          business.offboard: '#/components/schemas/BusinessOffboardProposalRequest'
    ProposalRequestBase:
      type: object
      properties:
        businessId:
          type: [string, 'null']
          description: The workspace the action applies to. Null only for practice-level actions.
    RouteProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.route }
            payload: { $ref: '#/components/schemas/RoutePayload' }
    UpdateCodingProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.update-coding }
            payload: { $ref: '#/components/schemas/UpdateCodingPayload' }
    MoveBusinessProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.move-business }
            payload: { $ref: '#/components/schemas/MoveBusinessPayload' }
    ReprocessProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.reprocess }
            payload: { $ref: '#/components/schemas/ReprocessPayload' }
    RejectProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.reject }
            payload: { $ref: '#/components/schemas/RejectPayload' }
    SplitProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.split }
            payload: { $ref: '#/components/schemas/SplitPayload' }
    ArchiveProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.archive }
            payload: { $ref: '#/components/schemas/ArchivePayload' }
    ChaseSendProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: chase.send }
            payload: { $ref: '#/components/schemas/ChaseSendPayload' }
    PublishBatchProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: publish.batch }
            payload: { $ref: '#/components/schemas/PublishBatchPayload' }
    BankConfirmMatchProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: bank.confirm-match }
            payload: { $ref: '#/components/schemas/BankConfirmMatchPayload' }
    RuleCreateProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: rule.create }
            payload: { $ref: '#/components/schemas/RuleCreatePayload' }
    RevokeLinkProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: document.revoke-link }
            payload: { $ref: '#/components/schemas/RevokeLinkPayload' }
    BusinessOffboardProposalRequest:
      allOf:
        - $ref: '#/components/schemas/ProposalRequestBase'
        - type: object
          required: [kind, payload]
          properties:
            kind: { const: business.offboard }
            payload: { $ref: '#/components/schemas/BusinessOffboardPayload' }
...
    ActionProposal:
      type: object
      required: [id, kind, state, payloadHash, createdAt, expiresAt]
      properties:
        id: { type: string }
        businessId: { type: [string, 'null'] }
        kind: { $ref: '#/components/schemas/ProposalKind' }
        state: { $ref: '#/components/schemas/ProposalState' }
        payload:
          type: object
          additionalProperties: true
          description: What was proposed, as submitted.
        payloadHash:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: SHA-256 over the canonical payload. Written to the audit event on approval.
        renderedSummaryHash:
          type: [string, 'null']
          description: Null until the review has been opened.
        createdByUserId: { type: [string, 'null'] }
        createdByModel:
          type: [string, 'null']
          description: |
            Set when a model proposed this. A model may create a proposal and
            may never approve one — the two fields are never both a model.
        createdAt: { type: string, format: date-time }
        reviewedAt:
          type: [string, 'null']
          format: date-time
          description: |
            Set by the review endpoint alone. While this is null the server
            refuses approval, whatever the UI shows.
        approvedByUserId: { type: [string, 'null'] }
        approvedAt: { type: [string, 'null'], format: date-time }
        executedAt: { type: [string, 'null'], format: date-time }
        expiresAt: { type: string, format: date-time }
        policyProposalId:
          type: [string, 'null']
          description: |
            Set on automated executions, naming the standing policy they ran
            under. A standing automation runs without a per-item proposal only
            because the policy itself was approved through this same contract
            (Governance §10.5).
        outcome:
          type: [object, 'null']
          additionalProperties: true
          description: What actually changed. Populated on execution.
        traceId: { type: [string, 'null'] }
...
    BankTransaction:
      type: object
      description: |
        The provider-agnostic normalised schema (SoT Stage 7). Amounts are
        signed integer pence exactly as the feed records them.
      required: [id, businessId, accountId, bookedAt, amountPence, currency, descriptionRaw, matchState, chaseSuppressed, matchedDocumentId]
      properties:
        id: { type: string }
        businessId: { type: string }
        accountId: { type: string }
        bookedAt: { type: string, format: date-time }
        amountPence:
          type: integer
          x-nt-money: true
          description: Signed. Negative is money out.
        matchedDocumentId:
          type: [string, 'null']
          description: |
            The document a CONFIRMED match links this line to, else null. The
            id the Bank screen's claimed-set and the approval drill-down key on
            — a SUGGESTED match deliberately does not fill it (a suggestion is
            a question, not evidence; `matchState` says one exists and the
            document's own bank-match read carries the detail).
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        descriptionRaw: { type: string }
        merchantName: { type: [string, 'null'] }
        classification: { type: [string, 'null'] }
        balanceAfterPence:
          type: [integer, 'null']
          x-nt-money: true
        matchState: { $ref: '#/components/schemas/MatchState' }
        chaseSuppressed:
          type: boolean
          description: |
            Bank-originated lines with no paperwork to chase — SERVICE CHARGE,
            STRIPE PAYOUT (SoT Stage 7). Suppressed lines never enter chase
            detection; the flag is visible so "why isn't this chased" has an
            answer on the screen.

    # ---- businesses (METH Stage 2, issue #120) ------------------------------
...